+++ include/pbvp/ui_bridge.hpp
@@ -50,6 +50,7 @@ public:
     static UiBridge& Instance() noexcept;
 
     void UpdateOnGameThread() noexcept;
+    [[nodiscard]] bool SetLayerEnabled(bool enabled) noexcept;
     [[nodiscard]] bool SetPlaybackStatus(
         const PlaybackStateSnapshot& playback) noexcept;
     UiRectSnapshot ReadForRenderThread() const noexcept;
@@ -74,6 +75,8 @@ private:
     bool map_visible_logged_{};
     std::uint32_t last_failure_{};
     std::uintptr_t last_status_tile_{};
+    std::uintptr_t last_root_tile_{};
+    bool last_layer_enabled_{true};
     PlaybackState last_status_state_{PlaybackState::unavailable};
     PlaybackError last_status_error_{PlaybackError::none};
     bool found_logged_{};